solid angle wrapper functional

+++ src/misc/solidangle.hh
@@ -1,9 +1,9 @@
 // solidangle.hh
 // header for solidangle.cpp
 
-void fastSAC(int num_pixels, float * theta, float * correction_factor);
+void fastSAC(int num_pixels, double * theta, double * correction_factor);
 
-void rigorousGridSAC(int num_pixels_s, int num_pixels_f, float * s,
-    float * f, float * p, float * correction_factor);
+void rigorousGridSAC(int num_pixels_s, int num_pixels_f, double * s,
+    double * f, double * p, double * correction_factor);
                      
-void rigorousExplicitSAC(float * pixel_xyz, float * correction_factor);
+void rigorousExplicitSAC(double * pixel_xyz, double * correction_factor);
